Move from_url and *_meta arguments to TimeSeries
Nothing about these is specific to Scenario objects/model data.
